Restructure crate to work on stable
Use proc_macro_hack crate in order to work on stable branches.

Add trybuild crate to better test for expected compile failures.
