+++ tests/cases/nul-compile-fail-1.rs
@@ -1,0 +1,5 @@
+use wchar::wch_c;
+
+fn main() {
+    wch_c!("fails \0 to compile");
+}
