+++ tests/cases/nul-compile-fail-2.rs
@@ -1,0 +1,5 @@
+use wchar::wch_c;
+
+fn main() {
+    wch_c!("also fails \0");
+}
