+++ wchar-impl/Cargo.toml
@@ -1,9 +1,8 @@
 [package]
-name = "wchar-macro"
-version = "0.1.0"
+name = "wchar-impl"
+version = "0.4.0"
 authors = ["Juici <juicy66173@gmail.com>"]
 description = "Procedural macros for compile time UTF-16 strings"
-
 edition = "2018"
 
 documentation = "https://docs.rs/wchar"
@@ -15,8 +14,6 @@ license = "MIT"
 keywords = ["wide", "string", "wchar", "utf16"]
 categories = ["text-processing"]
 
-include = ["Cargo.toml", "src/**/*.rs", "README.md", "LICENSE"]
-
 [badges]
 appveyor = { repository = "Juici/wchar-rs" }
 travis-ci = { repository = "Juici/wchar-rs" }
@@ -27,10 +24,6 @@ proc-macro = true
 
 [dependencies]
 proc-macro2 = "0.4"
-quote = "0.6"
 proc-macro-hack = "0.5"
-
-[dependencies.syn]
-version = "0.15"
-default-features = false
-features = ["derive", "parsing"]
+quote = "0.6"
+syn = { version = "0.15", features = ["full"] }
