--- wchar-rs/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "wchar"
-version = "0.3.0"
-authors = ["Juici <juicy66173@gmail.com>"]
-description = "Procedural macros for compile time UTF-16 strings"
-
-edition = "2018"
-
-documentation = "https://docs.rs/wchar"
-homepage = "https://github.com/Juici/wchar-rs"
-repository = "https://github.com/Juici/wchar-rs"
-readme = "README.md"
-license = "MIT"
-
-keywords = ["wide", "string", "wchar", "utf16"]
-categories = ["text-processing"]
-
-include = ["Cargo.toml", "src/**/*.rs", "README.md", "LICENSE"]
-
-[badges]
-appveyor = { repository = "Juici/wchar-rs" }
-travis-ci = { repository = "Juici/wchar-rs" }
-maintenance = { status = "passively-maintained" }
-
-[dependencies]
-wchar-macro = { path = "../wchar-rs-macro" }
-proc-macro-hack = "0.5"
